refactor and clean up diagnostic engine code so that it can be re-used stand alone.
code refactored is related to properly set up CompilationWithAnalyzer and compute and return diagnostics
