+++ src/Features/Core/Portable/Diagnostics/AnalyzerHelper.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -11,7 +12,11 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Diagnostics.Log;
 using Microsoft.CodeAnalysis.Diagnostics.Telemetry;
 using Microsoft.CodeAnalysis.ErrorReporting;
+using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.Options;
+using Microsoft.CodeAnalysis.PooledObjects;
+using Microsoft.CodeAnalysis.Shared.Extensions;
+using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Diagnostics
@@ -290,6 +295,465 @@ namespace Microsoft.CodeAnalysis.Diagnostics
         private static IEnumerable<AnalyzerPerformanceInfo> Convert(IEnumerable<(DiagnosticAnalyzer analyzer, TimeSpan timeSpan)> analyzerPerf, IDiagnosticAnalyzerService serviceOpt = null)
         {
             return analyzerPerf.Select(kv => new AnalyzerPerformanceInfo(kv.analyzer.GetAnalyzerId(), DiagnosticAnalyzerLogger.AllowsTelemetry(kv.analyzer, serviceOpt), kv.timeSpan));
+        }
+
+        /// <summary>
+        /// Get diagnostics for the given analyzers for the given document.
+        /// 
+        /// this is a simple API to get all diagnostics for the given document.
+        /// 
+        /// the intended audience for this API is for ones that pefer simplicity over performance such as document that belong to misc project.
+        /// this doesn't cache nor use cache for anything. it will re-caculate new diagnostics every time for the given document.
+        /// it will not persist any data on disk nor use OOP to calcuate the data.
+        /// 
+        /// this should never be used when performance is a big concern. for such context, use much complex API from IDiagnosticAnalyzerService
+        /// that provide all kinds of knobs/cache/persistency/OOP to get better perf over simplicity
+        /// </summary>
+        public static async Task<ImmutableArray<DiagnosticData>> GetDiagnosticsAsync(
+           this IDiagnosticAnalyzerService service, Document document, IEnumerable<DiagnosticAnalyzer> analyzers, AnalysisKind kind, CancellationToken cancellationToken)
+        {
+            // given service must be DiagnosticAnalyzerService
+            var diagnosticService = (DiagnosticAnalyzerService)service;
+
+            var analyzerDriverOpt = await diagnosticService.CreateAnalyzerDriverAsync(
+                document.Project, analyzers, includeSuppressedDiagnostics: false, logAggregatorOpt: null, cancellationToken).ConfigureAwait(false);
+
+            var builder = ArrayBuilder<DiagnosticData>.GetInstance();
+            foreach (var analyzer in analyzers)
+            {
+                builder.AddRange(await diagnosticService.ComputeDiagnosticsAsync(
+                    analyzerDriverOpt, document, analyzer, kind, spanOpt: null, logAggregatorOpt: null, cancellationToken).ConfigureAwait(false));
+            }
+
+            return builder.ToImmutableAndFree();
+        }
+
+        public static async Task<CompilationWithAnalyzers> CreateAnalyzerDriverAsync(
+                this DiagnosticAnalyzerService service,
+                Project project,
+                IEnumerable<DiagnosticAnalyzer> analyzers,
+                bool includeSuppressedDiagnostics,
+                DiagnosticLogAggregator logAggregatorOpt,
+                CancellationToken cancellationToken)
+        {
+            if (!project.SupportsCompilation)
+            {
+                return null;
+            }
+
+            var compilation = await project.GetCompilationAsync(cancellationToken).ConfigureAwait(false);
+
+            // Create driver that holds onto compilation and associated analyzers
+            var filteredAnalyzers = analyzers.Where(a => !a.IsWorkspaceDiagnosticAnalyzer()).ToImmutableArrayOrEmpty();
+
+            // PERF: there is no analyzers for this compilation.
+            //       compilationWithAnalyzer will throw if it is created with no analyzers which is perf optimization.
+            if (filteredAnalyzers.IsEmpty)
+            {
+                return null;
+            }
+
+            Contract.ThrowIfFalse(project.SupportsCompilation);
+            AssertCompilation(project, compilation);
+
+            // Create driver that holds onto compilation and associated analyzers
+            return compilation.WithAnalyzers(filteredAnalyzers, GetAnalyzerOptions());
+
+            CompilationWithAnalyzersOptions GetAnalyzerOptions()
+            {
+                // in IDE, we always set concurrentAnalysis == false otherwise, we can get into thread starvation due to
+                // async being used with syncronous blocking concurrency.
+                return new CompilationWithAnalyzersOptions(
+                    options: new WorkspaceAnalyzerOptions(project.AnalyzerOptions, project.Solution.Options, project.Solution),
+                    onAnalyzerException: service.GetOnAnalyzerException(project.Id, logAggregatorOpt),
+                    analyzerExceptionFilter: GetAnalyzerExceptionFilter(),
+                    concurrentAnalysis: false,
+                    logAnalyzerExecutionTime: true,
+                    reportSuppressedDiagnostics: includeSuppressedDiagnostics);
+            }
+
+            Func<Exception, bool> GetAnalyzerExceptionFilter()
+            {
+                return ex =>
+                {
+                    if (project.Solution.Workspace.Options.GetOption(InternalDiagnosticsOptions.CrashOnAnalyzerException))
+                    {
+                        // if option is on, crash the host to get crash dump.
+                        FatalError.ReportUnlessCanceled(ex);
+                    }
+
+                    return true;
+                };
+            }
+        }
+
+        [Conditional("DEBUG")]
+        private static void AssertCompilation(Project project, Compilation compilation1)
+        {
+            // given compilation must be from given project.
+            Contract.ThrowIfFalse(project.TryGetCompilation(out var compilation2));
+            Contract.ThrowIfFalse(compilation1 == compilation2);
+        }
+
+        /// <summary>
+        /// Return all local diagnostics (syntax, semantic) that belong to given document for the given StateSet (analyzer) by calculating them
+        /// </summary>
+        public static async Task<IEnumerable<DiagnosticData>> ComputeDiagnosticsAsync(
+            this DiagnosticAnalyzerService service,
+            CompilationWithAnalyzers analyzerDriverOpt,
+            Document document,
+            DiagnosticAnalyzer analyzer,
+            AnalysisKind kind,
+            TextSpan? spanOpt,
+            DiagnosticLogAggregator logAggregatorOpt,
+            CancellationToken cancellationToken)
+        {
+            if (analyzer is DocumentDiagnosticAnalyzer documentAnalyzer)
+            {
+                var diagnostics = await ComputeDocumentDiagnosticAnalyzerDiagnosticsAsync(
+                    service, document, documentAnalyzer, kind, analyzerDriverOpt?.Compilation, logAggregatorOpt, cancellationToken).ConfigureAwait(false);
+                return diagnostics.ConvertToLocalDiagnostics(document);
+            }
+
+            var documentDiagnostics = await ComputeDiagnosticAnalyzerDiagnosticsAsync().ConfigureAwait(false);
+
+            return documentDiagnostics.ConvertToLocalDiagnostics(document);
+
+            async Task<IEnumerable<Diagnostic>> ComputeDiagnosticAnalyzerDiagnosticsAsync()
+            {
+                // quick optimization to reduce allocations.
+                if (analyzerDriverOpt == null || !service.SupportAnalysisKind(analyzer, document.Project.Language, kind))
+                {
+                    LogSyntaxInfoWithoutDiagnostics();
+                    return ImmutableArray<Diagnostic>.Empty;
+                }
+
+                if (!await AnalysisEnabled().ConfigureAwait(false))
+                {
+                    return ImmutableArray<Diagnostic>.Empty;
+                }
+
+                // REVIEW: more unnecessary allocations just to get diagnostics per analyzer
+                var oneAnalyzers = ImmutableArray.Create(analyzer);
+
+                switch (kind)
+                {
+                    case AnalysisKind.Syntax:
+                        var tree = await document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
+                        var diagnostics = await analyzerDriverOpt.GetAnalyzerSyntaxDiagnosticsAsync(tree, oneAnalyzers, cancellationToken).ConfigureAwait(false);
+                        LogSyntaxInfo(diagnostics, tree);
+
+                        Debug.Assert(diagnostics.Count() == CompilationWithAnalyzers.GetEffectiveDiagnostics(diagnostics, analyzerDriverOpt.Compilation).Count());
+                        return diagnostics.ToImmutableArrayOrEmpty();
+                    case AnalysisKind.Semantic:
+                        var model = await document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
+                        diagnostics = await analyzerDriverOpt.GetAnalyzerSemanticDiagnosticsAsync(model, spanOpt, oneAnalyzers, cancellationToken).ConfigureAwait(false);
+
+                        Debug.Assert(diagnostics.Count() == CompilationWithAnalyzers.GetEffectiveDiagnostics(diagnostics, analyzerDriverOpt.Compilation).Count());
+                        return diagnostics.ToImmutableArrayOrEmpty();
+                    default:
+                        return Contract.FailWithReturn<ImmutableArray<Diagnostic>>("shouldn't reach here");
+                }
+            }
+
+            async Task<bool> AnalysisEnabled()
+            {
+                // if project is not loaded successfully then, we disable semantic errors for compiler analyzers
+                if (kind == AnalysisKind.Syntax || !service.IsCompilerDiagnosticAnalyzer(document.Project.Language, analyzer))
+                {
+                    return true;
+                }
+
+                var enabled = await document.Project.HasSuccessfullyLoadedAsync(cancellationToken).ConfigureAwait(false);
+
+                Logger.Log(FunctionId.Diagnostics_SemanticDiagnostic, (a, d, e) => $"{a.ToString()}, ({d.Id}, {d.Project.Id}), Enabled:{e}", analyzer, document, enabled);
+
+                return enabled;
+            }
+
+            void LogSyntaxInfo(ImmutableArray<Diagnostic> diagnostics, SyntaxTree tree)
+            {
+                if (!diagnostics.IsDefaultOrEmpty)
+                {
+                    return;
+                }
+
+                Logger.Log(FunctionId.Diagnostics_SyntaxDiagnostic,
+                    (d, a, t) => $"{d.Id}, {d.Project.Id}, {a.ToString()}, {t.Length}", document, analyzer, tree);
+            }
+
+            void LogSyntaxInfoWithoutDiagnostics()
+            {
+                if (kind != AnalysisKind.Syntax)
+                {
+                    return;
+                }
+
+                Logger.Log(FunctionId.Diagnostics_SyntaxDiagnostic,
+                    (r, d, a, k) => $"Driver: {r != null}, {d.Id}, {d.Project.Id}, {a.ToString()}, {k}", analyzerDriverOpt, document, analyzer, kind);
+            }
+        }
+
+        public static bool SupportAnalysisKind(this DiagnosticAnalyzerService service, DiagnosticAnalyzer analyzer, string language, AnalysisKind kind)
+        {
+            // compiler diagnostic analyzer always support all kinds
+            if (service.IsCompilerDiagnosticAnalyzer(language, analyzer))
+            {
+                return true;
+            }
+
+            switch (kind)
+            {
+                case AnalysisKind.Syntax:
+                    return analyzer.SupportsSyntaxDiagnosticAnalysis();
+                case AnalysisKind.Semantic:
+                    return analyzer.SupportsSemanticDiagnosticAnalysis();
+                default:
+                    return Contract.FailWithReturn<bool>("shouldn't reach here");
+            }
+        }
+
+        public static async Task<IEnumerable<Diagnostic>> ComputeDocumentDiagnosticAnalyzerDiagnosticsAsync(
+                this DiagnosticAnalyzerService service,
+                Document document,
+                DocumentDiagnosticAnalyzer analyzer,
+                AnalysisKind kind,
+                Compilation compilationOpt,
+                DiagnosticLogAggregator logAggregatorOpt,
+                CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+
+            try
+            {
+                Task<ImmutableArray<Diagnostic>> analyzeAsync;
+
+                switch (kind)
+                {
+                    case AnalysisKind.Syntax:
+                        analyzeAsync = analyzer.AnalyzeSyntaxAsync(document, cancellationToken);
+                        break;
+
+                    case AnalysisKind.Semantic:
+                        analyzeAsync = analyzer.AnalyzeSemanticsAsync(document, cancellationToken);
+                        break;
+
+                    default:
+                        throw ExceptionUtilities.UnexpectedValue(kind);
+                }
+
+                var diagnostics = (await analyzeAsync.ConfigureAwait(false)).NullToEmpty();
+                if (compilationOpt != null)
+                {
+                    return CompilationWithAnalyzers.GetEffectiveDiagnostics(diagnostics, compilationOpt);
+                }
+
+#if DEBUG
+                // since all DocumentDiagnosticAnalyzers are from internal users, we only do debug check. also this can be expensive at runtime
+                // since it requires await. if we find any offender through NFW, we should be able to fix those since all those should
+                // from intern teams.
+                await VerifyDiagnosticLocationsAsync(diagnostics, document.Project, cancellationToken).ConfigureAwait(false);
+#endif
+
+                return diagnostics;
+            }
+            catch (Exception e) when (!IsCanceled(e, cancellationToken))
+            {
+                OnAnalyzerException(service, analyzer, document.Project.Id, compilationOpt, e, logAggregatorOpt);
+                return ImmutableArray<Diagnostic>.Empty;
+            }
+        }
+
+        public static async Task<IEnumerable<Diagnostic>> ComputeProjectDiagnosticAnalyzerDiagnosticsAsync(
+                this DiagnosticAnalyzerService service,
+                Project project,
+                ProjectDiagnosticAnalyzer analyzer,
+                Compilation compilationOpt,
+                DiagnosticLogAggregator logAggregatorOpt,
+                CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+
+            try
+            {
+                var diagnostics = (await analyzer.AnalyzeProjectAsync(project, cancellationToken).ConfigureAwait(false)).NullToEmpty();
+
+                // Apply filtering from compilation options (source suppressions, ruleset, etc.)
+                if (compilationOpt != null)
+                {
+                    diagnostics = CompilationWithAnalyzers.GetEffectiveDiagnostics(diagnostics, compilationOpt).ToImmutableArrayOrEmpty();
+                }
+
+#if DEBUG
+                // since all ProjectDiagnosticAnalyzers are from internal users, we only do debug check. also this can be expensive at runtime
+                // since it requires await. if we find any offender through NFW, we should be able to fix those since all those should
+                // from intern teams.
+                await VerifyDiagnosticLocationsAsync(diagnostics, project, cancellationToken).ConfigureAwait(false);
+#endif
+
+                return diagnostics;
+            }
+            catch (Exception e) when (!IsCanceled(e, cancellationToken))
+            {
+                OnAnalyzerException(service, analyzer, project.Id, compilationOpt, e, logAggregatorOpt);
+                return ImmutableArray<Diagnostic>.Empty;
+            }
+        }
+
+        private static bool IsCanceled(Exception ex, CancellationToken cancellationToken)
+        {
+            return (ex as OperationCanceledException)?.CancellationToken == cancellationToken;
+        }
+
+        private static void OnAnalyzerException(
+            DiagnosticAnalyzerService service, DiagnosticAnalyzer analyzer, ProjectId projectId, Compilation compilationOpt, Exception ex, DiagnosticLogAggregator logAggregatorOpt)
+        {
+            var exceptionDiagnostic = AnalyzerHelper.CreateAnalyzerExceptionDiagnostic(analyzer, ex);
+
+            if (compilationOpt != null)
+            {
+                exceptionDiagnostic = CompilationWithAnalyzers.GetEffectiveDiagnostics(ImmutableArray.Create(exceptionDiagnostic), compilationOpt).SingleOrDefault();
+            }
+
+            var onAnalyzerException = service.GetOnAnalyzerException(projectId, logAggregatorOpt);
+            onAnalyzerException?.Invoke(ex, analyzer, exceptionDiagnostic);
+        }
+
+        private static async Task VerifyDiagnosticLocationsAsync(ImmutableArray<Diagnostic> diagnostics, Project project, CancellationToken cancellationToken)
+        {
+            foreach (var diagnostic in diagnostics)
+            {
+                await VerifyDiagnosticLocationAsync(diagnostic.Id, diagnostic.Location).ConfigureAwait(false);
+
+                if (diagnostic.AdditionalLocations != null)
+                {
+                    foreach (var location in diagnostic.AdditionalLocations)
+                    {
+                        await VerifyDiagnosticLocationAsync(diagnostic.Id, diagnostic.Location).ConfigureAwait(false);
+                    }
+                }
+            }
+
+            async Task VerifyDiagnosticLocationAsync(string id, Location location)
+            {
+                switch (location.Kind)
+                {
+                    case LocationKind.None:
+                    case LocationKind.MetadataFile:
+                    case LocationKind.XmlFile:
+                        // ignore these kinds
+                        break;
+                    case LocationKind.SourceFile:
+                    {
+                        if (project.GetDocument(location.SourceTree) == null)
+                        {
+                            // Disallow diagnostics with source locations outside this project.
+                            throw new ArgumentException(string.Format(FeaturesResources.Reported_diagnostic_0_has_a_source_location_in_file_1_which_is_not_part_of_the_compilation_being_analyzed, id, location.SourceTree.FilePath), "diagnostic");
+                        }
+
+                        if (location.SourceSpan.End > location.SourceTree.Length)
+                        {
+                            // Disallow diagnostics with source locations outside this project.
+                            throw new ArgumentException(string.Format(FeaturesResources.Reported_diagnostic_0_has_a_source_location_1_in_file_2_which_is_outside_of_the_given_file, id, location.SourceSpan, location.SourceTree.FilePath), "diagnostic");
+                        }
+                    }
+                    break;
+                    case LocationKind.ExternalFile:
+                    {
+                        var filePath = location.GetLineSpan().Path;
+                        var document = TryGetDocumentWithFilePath(filePath);
+                        if (document == null)
+                        {
+                            // this is not a roslyn file. we don't care about this file.
+                            return;
+                        }
+
+                        // this can be potentially expensive since it will load text if it is not already loaded.
+                        // but, this text is most likely already loaded since producer of this diagnostic (Document/ProjectDiagnosticAnalyzers)
+                        // should have loaded it to produce the diagnostic at the first place. once loaded, it should stay in memory until
+                        // project cache goes away. when text is already there, await should return right away.
+                        var text = await document.GetTextAsync(cancellationToken).ConfigureAwait(false);
+                        if (location.SourceSpan.End > text.Length)
+                        {
+                            // Disallow diagnostics with locations outside this project.
+                            throw new ArgumentException(string.Format(FeaturesResources.Reported_diagnostic_0_has_a_source_location_1_in_file_2_which_is_outside_of_the_given_file, id, location.SourceSpan, filePath), "diagnostic");
+                        }
+                    }
+                    break;
+                    default:
+                        throw ExceptionUtilities.Unreachable;
+                }
+            }
+
+            Document TryGetDocumentWithFilePath(string path)
+            {
+                foreach (var documentId in project.Solution.GetDocumentIdsWithFilePath(path))
+                {
+                    if (documentId.ProjectId == project.Id)
+                    {
+                        return project.GetDocument(documentId);
+                    }
+                }
+
+                return null;
+            }
+        }
+
+        public static IEnumerable<DiagnosticData> ConvertToLocalDiagnostics(this IEnumerable<Diagnostic> diagnostics, Document targetDocument, TextSpan? span = null)
+        {
+            var project = targetDocument.Project;
+
+            if (project.SupportsCompilation)
+            {
+                return ConvertToLocalDiagnosticsWithCompilation();
+            }
+
+            return ConvertToLocalDiagnosticsWithoutCompilation();
+
+            IEnumerable<DiagnosticData> ConvertToLocalDiagnosticsWithoutCompilation()
+            {
+                Contract.ThrowIfTrue(project.SupportsCompilation);
+
+                foreach (var diagnostic in diagnostics)
+                {
+                    var location = diagnostic.Location;
+                    if (location.Kind != LocationKind.ExternalFile)
+                    {
+                        continue;
+                    }
+
+                    var lineSpan = location.GetLineSpan();
+
+                    var documentIds = project.Solution.GetDocumentIdsWithFilePath(lineSpan.Path);
+                    if (documentIds.IsEmpty || documentIds.All(id => id != targetDocument.Id))
+                    {
+                        continue;
+                    }
+
+                    yield return DiagnosticData.Create(targetDocument, diagnostic);
+                }
+            }
+
+            IEnumerable<DiagnosticData> ConvertToLocalDiagnosticsWithCompilation()
+            {
+                Contract.ThrowIfFalse(project.SupportsCompilation);
+
+                foreach (var diagnostic in diagnostics)
+                {
+                    var document = project.GetDocument(diagnostic.Location.SourceTree);
+                    if (document == null || document != targetDocument)
+                    {
+                        continue;
+                    }
+
+                    if (span.HasValue && !span.Value.IntersectsWith(diagnostic.Location.SourceSpan))
+                    {
+                        continue;
+                    }
+
+                    yield return DiagnosticData.Create(document, diagnostic);
+                }
+            }
         }
 
         public static bool IsCompilationEndAnalyzer(this DiagnosticAnalyzer analyzer, Project project, Compilation compilation)
