+++ src/Features/Core/Portable/Diagnostics/EngineV2/AnalysisKind.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Diagnostics
+{
+    /// <summary>
+    /// enum for each analysis kind.
+    /// </summary>
+    internal enum AnalysisKind
+    {
+        Syntax,
+        Semantic,
+        NonLocal
+    }
+}
