+++ src/Features/Core/Portable/Diagnostics/EngineV2/DiagnosticIncrementalAnalyzer.CompilationManager.cs
@@ -1,14 +1,11 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.ErrorReporting;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Diagnostics.EngineV2
@@ -70,79 +67,10 @@ namespace Microsoft.CodeAnalysis.Diagnostics.EngineV2
                 return CreateAnalyzerDriverAsync(project, analyzers, includeSuppressedDiagnostics, cancellationToken);
             }
 
-            public async Task<CompilationWithAnalyzers> CreateAnalyzerDriverAsync(
+            public Task<CompilationWithAnalyzers> CreateAnalyzerDriverAsync(
                 Project project, IEnumerable<DiagnosticAnalyzer> analyzers, bool includeSuppressedDiagnostics, CancellationToken cancellationToken)
             {
-                if (!project.SupportsCompilation)
-                {
-                    return null;
-                }
-
-                var compilation = await project.GetCompilationAsync(cancellationToken).ConfigureAwait(false);
-
-                // Create driver that holds onto compilation and associated analyzers
-                return CreateAnalyzerDriver(
-                    project, compilation, analyzers, logAnalyzerExecutionTime: true, reportSuppressedDiagnostics: includeSuppressedDiagnostics);
-            }
-
-            private CompilationWithAnalyzers CreateAnalyzerDriver(
-                Project project,
-                Compilation compilation,
-                IEnumerable<DiagnosticAnalyzer> allAnalyzers,
-                bool logAnalyzerExecutionTime,
-                bool reportSuppressedDiagnostics)
-            {
-                var analyzers = allAnalyzers.Where(a => !a.IsWorkspaceDiagnosticAnalyzer()).ToImmutableArrayOrEmpty();
-
-                // PERF: there is no analyzers for this compilation.
-                //       compilationWithAnalyzer will throw if it is created with no analyzers which is perf optimization.
-                if (analyzers.IsEmpty)
-                {
-                    return null;
-                }
-
-                Contract.ThrowIfFalse(project.SupportsCompilation);
-                AssertCompilation(project, compilation);
-
-                var analysisOptions = GetAnalyzerOptions(project, logAnalyzerExecutionTime, reportSuppressedDiagnostics);
-
-                // Create driver that holds onto compilation and associated analyzers
-                return compilation.WithAnalyzers(analyzers, analysisOptions);
-            }
-
-            private CompilationWithAnalyzersOptions GetAnalyzerOptions(
-                Project project,
-                bool logAnalyzerExecutionTime,
-                bool reportSuppressedDiagnostics)
-            {
-                // in IDE, we always set concurrentAnalysis == false otherwise, we can get into thread starvation due to
-                // async being used with syncronous blocking concurrency.
-                return new CompilationWithAnalyzersOptions(
-                    options: new WorkspaceAnalyzerOptions(project.AnalyzerOptions, project.Solution.Options, project.Solution),
-                    onAnalyzerException: GetOnAnalyzerException(project.Id),
-                    analyzerExceptionFilter: GetAnalyzerExceptionFilter(project),
-                    concurrentAnalysis: false,
-                    logAnalyzerExecutionTime: logAnalyzerExecutionTime,
-                    reportSuppressedDiagnostics: reportSuppressedDiagnostics);
-            }
-
-            private Func<Exception, bool> GetAnalyzerExceptionFilter(Project project)
-            {
-                return ex =>
-                {
-                    if (project.Solution.Workspace.Options.GetOption(InternalDiagnosticsOptions.CrashOnAnalyzerException))
-                    {
-                        // if option is on, crash the host to get crash dump.
-                        FatalError.ReportUnlessCanceled(ex);
-                    }
-
-                    return true;
-                };
-            }
-
-            private Action<Exception, DiagnosticAnalyzer, Diagnostic> GetOnAnalyzerException(ProjectId projectId)
-            {
-                return _owner.Owner.GetOnAnalyzerException(projectId, _owner.DiagnosticLogAggregator);
+                return _owner.Owner.CreateAnalyzerDriverAsync(project, analyzers, includeSuppressedDiagnostics, _owner.DiagnosticLogAggregator, cancellationToken);
             }
 
             private void ResetAnalyzerDriverMap()
@@ -167,14 +95,6 @@ namespace Microsoft.CodeAnalysis.Diagnostics.EngineV2
 
                 // make sure analyzers are same.
                 Contract.ThrowIfFalse(analyzerDriver.Analyzers.SetEquals(stateSets.Select(s => s.Analyzer).Where(a => !a.IsWorkspaceDiagnosticAnalyzer())));
-            }
-
-            [Conditional("DEBUG")]
-            private void AssertCompilation(Project project, Compilation compilation1)
-            {
-                // given compilation must be from given project.
-                Contract.ThrowIfFalse(project.TryGetCompilation(out var compilation2));
-                Contract.ThrowIfFalse(compilation1 == compilation2);
             }
 
             #region state changed
